Move files for 0.88
