+++ custom_updater.json
@@ -1,26 +1,26 @@
 {
    "device_tracker.hue" : {
-      "updated_at" : "2019-02-06",
-      "local_location" : "/custom_components/device_tracker/hue.py",
-      "version" : "1.0.4",
+      "updated_at" : "2019-02-24",
+      "local_location" : "/custom_components/hue/device_tracker.py",
+      "version" : "1.0.5",
       "changelog" : "https://github.com/robmarkcole/Hue-sensors-HASS/releases/latest",
       "visit_repo" : "https://github.com/robmarkcole/Hue-sensors-HASS",
-      "remote_location" : "https://raw.githubusercontent.com/robmarkcole/Hue-sensors-HASS/master/custom_components/device_tracker/hue.py"
+      "remote_location" : "https://raw.githubusercontent.com/robmarkcole/Hue-sensors-HASS/master/custom_components/hue/device_tracker.py"
    },
    "binary_sensor.hue" : {
       "changelog" : "https://github.com/robmarkcole/Hue-sensors-HASS/releases/latest",
-      "version" : "1.0.4",
-      "local_location" : "/custom_components/binary_sensor/hue.py",
-      "updated_at" : "2019-02-06",
+      "version" : "1.0.5",
+      "local_location" : "/custom_components/hue/binary_sensor.py",
+      "updated_at" : "2019-02-24",
       "visit_repo" : "https://github.com/robmarkcole/Hue-sensors-HASS",
-      "remote_location" : "https://raw.githubusercontent.com/robmarkcole/Hue-sensors-HASS/master/custom_components/binary_sensor/hue.py"
+      "remote_location" : "https://raw.githubusercontent.com/robmarkcole/Hue-sensors-HASS/master/custom_components/hue/binary_sensor.py"
    },
    "sensor.hue" : {
-      "version" : "1.0.4",
+      "version" : "1.0.5",
       "changelog" : "https://github.com/robmarkcole/Hue-sensors-HASS/releases/latest",
-      "updated_at" : "2019-02-06",
-      "local_location" : "/custom_components/sensor/hue.py",
-      "remote_location" : "https://raw.githubusercontent.com/robmarkcole/Hue-sensors-HASS/master/custom_components/sensor/hue.py",
+      "updated_at" : "2019-02-24",
+      "local_location" : "/custom_components/hue/sensor.py",
+      "remote_location" : "https://raw.githubusercontent.com/robmarkcole/Hue-sensors-HASS/master/custom_components/hue/sensor.py",
       "visit_repo" : "https://github.com/robmarkcole/Hue-sensors-HASS"
    }
 }
